requirements.txt
discord.py>=2.3.0
matplotlib>=3.8.0
tzdata>=2024.1
sqlite-web>=0.6.0
Pillow>=10.0.0
